requirements.txt
# Core dependencies
python-dotenv # For environment variables
pygame # For music player functionality
pynput # For keyboard input handling
csv # Built-in, for CSV file operations
pillow # For image processing
colorama  # For better color support across platforms
flask # For web server functionality
flask-login # For user authentication
werkzeug # For utilities including password hashing

# Built-in modules used
# datetime  - For date and time operations
# tkinter   - For GUI development
# shelve    - For simple database operations
# os        - For file and system operations
# time      - For delays and timing
# csv       - For CSV file operations
# random    - For random number generation
# sys       - For system operations
# hashlib   - For hashing operations
# hmac      - For keyed-hashing for message authentication